fastapi_dana_SOA/requirements.txt
fastapi
uvicorn
hypercorn[h2]    # Support HTTP/2.0 [cite: 17]
sqlalchemy
aiomysql         # Driver MySQL Async (not used by auth now)
pymysql          # Driver MySQL untuk auth-service (SOAP DB)
pydantic
pydantic-settings
python-jose[cryptography] # Untuk JWT [cite: 37]
passlib[bcrypt]           # (tidak digunakan, bisa dihapus nanti)
aiokafka                  # Kafka client async [cite: 43]
locust                    # Untuk load testing [cite: 26]
python-dotenv
cryptography